Move PrivateReposDashboard into dashboards module
